refresh.py: generate a `compile_commond.json` using `arguments` instead of `command`
In case the generated command-line contains whitespaces, it is easier to
use the command in the list-form that we have and generate an `argument`
directive instead of joining the list to be used for `command` and
possibly introducing escaping errors even though `shlex.join()` should
be safe.

By using `arguments` instead of `shlex.join()` we also keep
compatibility with Python 3.7.